Added patient required fields for syncing and sync validation status field
